feature: Prevent deletion of Integration Provider with Integration Apps (#165)

+++ packages/api/src/integration-providers-global/integration-providers-global.controller.ts
@@ -3,7 +3,7 @@ import {
   PostIntegrationProviderDto,
   toGetIntegrationProviderDto,
 } from '@edanalytics/models';
-import { IntegrationProvider, Ownership } from '@edanalytics/models-server';
+import { IntegrationAppDetailed, IntegrationProvider, Ownership } from '@edanalytics/models-server';
 
 import {
   Body,
@@ -29,9 +29,30 @@ export class IntegrationProvidersGlobalController {
   constructor(
     @InjectRepository(IntegrationProvider)
     private integrationProvidersRepository: Repository<IntegrationProvider>,
+    @InjectRepository(IntegrationAppDetailed)
+    private integrationAppsDetailedRepository: Repository<IntegrationAppDetailed>,
     @InjectRepository(Ownership)
     private ownershipsViewRepository: Repository<Ownership>
   ) {}
+
+  async getOneIntegrationAppCount(integrationProviderId: number) {
+    return await this.integrationAppsDetailedRepository.count({
+      where: { integrationProviderId },
+    });
+  }
+
+  async getManyIntegrationAppCounts() {
+    const appCounts = await this.integrationAppsDetailedRepository
+      .createQueryBuilder('app')
+      .select('app.integrationProviderId, COUNT(app.id) AS "appCount"')
+      .groupBy('app.integrationProviderId')
+      .getRawMany();
+
+    return appCounts.reduce((object, { integrationProviderId, appCount }) => {
+      object[integrationProviderId] = Number(appCount);
+      return object;
+    }, {});
+  }
 
   @Post()
   @Authorize({
@@ -58,7 +79,12 @@ export class IntegrationProvidersGlobalController {
   }
 
   async findOne(id: number) {
-    return await this.integrationProvidersRepository.findOneByOrFail({ id }).catch(throwNotFound);
+    const integrationProvider = await this.integrationProvidersRepository
+      .findOneByOrFail({ id })
+      .catch(throwNotFound);
+    const integrationAppCount = await this.getOneIntegrationAppCount(integrationProvider.id);
+    integrationProvider.appCount = integrationAppCount;
+    return integrationProvider;
   }
 
   @Get()
@@ -81,9 +107,13 @@ export class IntegrationProvidersGlobalController {
       );
     }
 
-    return toGetIntegrationProviderDto(
-      await this.integrationProvidersRepository.find().catch(throwNotFound)
-    );
+    const appCounts = await this.getManyIntegrationAppCounts();
+    const integrationProviders = await this.integrationProvidersRepository.find();
+    integrationProviders.forEach((provider) => {
+      provider.appCount = appCounts[provider.id] || 0;
+    });
+
+    return toGetIntegrationProviderDto(integrationProviders);
   }
 
   @Put(':id')
@@ -111,6 +141,13 @@ export class IntegrationProvidersGlobalController {
   })
   async remove(@Param('id', new ParseIntPipe()) id: number) {
     const old = await this.findOne(id).catch(throwNotFound);
+    const integrationAppCount = await this.getOneIntegrationAppCount(old.id);
+    if (integrationAppCount > 0) {
+      throw new HttpException(
+        'This Integration Provider has associated applications and cannot be deleted',
+        400
+      );
+    }
     await this.integrationProvidersRepository.remove(old);
     return undefined;
   }
